changed user to member, removed admin.guard
